+++ src/main/java/htw/calenderproject/persistence/event/EventRepository.java
@@ -1,4 +1,4 @@
-package htw.calenderproject;
+package htw.calenderproject.persistence.event;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
