--- src/main/java/htw/calenderproject/service/EventService.java
@@ -1,16 +1,0 @@
-package htw.calenderproject.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class EventService {
-
-    //current user
-    public static void setOwner(String userName) {
-
-    }
-
-    //add event to an existing calender
-    //remove an event from an existing calender
-    //list all events
-}
